backend/requirements.txt
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
psycopg[binary]==3.1.12
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2

